+++ tagger/src/components/image.rs
@@ -2,7 +2,6 @@ use crate::components::Title;
 
 use super::Render;
 use anyhow::Result;
-use image::ImageReader;
 use ratatui::{
     layout::{Constraint, Direction, Layout, Rect},
     Frame,
@@ -62,10 +61,7 @@ impl<'a> Image<'a> {
 impl Render for Image<'_> {
     fn render(&mut self, f: &mut Frame<'_>, area: Rect) -> Result<()> {
         fn inner(this: &mut Image<'_>, f: &mut Frame<'_>, area: Rect) -> Result<()> {
-            let dyn_img = match this.path.is_symlink() {
-                true => ImageReader::open(this.path.read_link()?)?.decode()?,
-                false => ImageReader::open(this.path)?.decode()?,
-            };
+            let dyn_img = image::open(this.path.canonicalize()?)?;
             let mut image_fit_state = this.picker.new_resize_protocol(dyn_img);
             let image = StatefulImage::new(None).resize(Resize::Fit(Some(FilterType::Gaussian)));
             f.render_stateful_widget(image, area, &mut image_fit_state);